Task 2.16: remove extra functions, fix templates format.
