+++ src/view/form-edit-event.js
@@ -1,15 +1,11 @@
-import { formatDate } from '../utils';
-import { TYPES } from '../mock/events';
+import { TYPES } from '../const';
 import { TOWNS } from '../mock/event-destination';
-import { renderTypes } from '../utils';
-import { renderDestinations } from '../utils';
-import { renderOffers } from '../utils';
+import { renderTypes, renderDestinations, renderOffers } from './view-utils';
 
 const createFormEditEventTemplate = (event) => {
   const { base_price: price, date_from: dateFrom, date_to: dateTo, destination, id, offers, type } = event;
   const {description, name} = destination;
-  const formEventTemplate = `
-	<li class="trip-events__item">
+  const formEventTemplate = `<li class="trip-events__item">
 		<form class="event event--edit" action="#" method="post">
 			<header class="event__header">
 				<div class="event__type-wrapper">
@@ -40,10 +36,10 @@ const createFormEditEventTemplate = (event) => {
 
 				<div class="event__field-group  event__field-group--time">
 					<label class="visually-hidden" for="event-start-time-${id}">From</label>
-					<input class="event__input  event__input--time" id="event-start-time-${id}" type="text" name="event-start-time" value="${formatDate(dateFrom, 'YY/MM/DD HH:MM')}">
+					<input class="event__input  event__input--time" id="event-start-time-${id}" type="text" name="event-start-time" value="${dateFrom.format('YY/MM/DD HH:MM')}">
 					&mdash;
 					<label class="visually-hidden" for="event-end-time-${id}">To</label>
-					<input class="event__input  event__input--time" id="event-end-time-${id}" type="text" name="event-end-time" value="${formatDate(dateTo, 'YY/MM/DD HH:MM')}">
+					<input class="event__input  event__input--time" id="event-end-time-${id}" type="text" name="event-end-time" value="${dateTo.format('YY/MM/DD HH:MM')}">
 				</div>
 
 				<div class="event__field-group  event__field-group--price">
