+++ src/view/menu.js
@@ -1,9 +1,7 @@
-export const createMenuTemplate = () => (`
-  <div class="trip-controls__navigation">
-    <h2 class="visually-hidden">Switch trip view</h2>
-    <nav class="trip-controls__trip-tabs  trip-tabs">
-      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-      <a class="trip-tabs__btn" href="#">Stats</a>
-    </nav>
-  </div>
-`);
+export const createMenuTemplate = () => (`<div class="trip-controls__navigation">
+  <h2 class="visually-hidden">Switch trip view</h2>
+  <nav class="trip-controls__trip-tabs  trip-tabs">
+    <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
+    <a class="trip-tabs__btn" href="#">Stats</a>
+  </nav>
+</div>`);
