+++ src/view/trip-filters.js
@@ -1,24 +1,22 @@
 
-export const createTripFiltersTemplate = () => (`
-	<div class="trip-controls__filters">
-    <h2 class="visually-hidden">Filter events</h2>
-    <form class="trip-filters" action="#" method="get">
-      <div class="trip-filters__filter">
-        <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked="">
-        <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
-      </div>
+export const createTripFiltersTemplate = () => (`<div class="trip-controls__filters">
+  <h2 class="visually-hidden">Filter events</h2>
+  <form class="trip-filters" action="#" method="get">
+    <div class="trip-filters__filter">
+      <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked="">
+      <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
+    </div>
 
-      <div class="trip-filters__filter">
-        <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
-        <label class="trip-filters__filter-label" for="filter-future">Future</label>
-      </div>
+    <div class="trip-filters__filter">
+      <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
+      <label class="trip-filters__filter-label" for="filter-future">Future</label>
+    </div>
 
-      <div class="trip-filters__filter">
-        <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
-        <label class="trip-filters__filter-label" for="filter-past">Past</label>
-      </div>
+    <div class="trip-filters__filter">
+      <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
+      <label class="trip-filters__filter-label" for="filter-past">Past</label>
+    </div>
 
-      <button class="visually-hidden" type="submit">Accept filter</button>
-    </form>
-  </div>
-`);
+    <button class="visually-hidden" type="submit">Accept filter</button>
+  </form>
+</div>`);
