+++ src/view/trip-info-section.js
@@ -1,4 +1,2 @@
-export const createTripInfoSectionTemplate = () => (`
-  <section class="trip-main__trip-info  trip-info">
-  </section>
-`);
+export const createTripInfoSectionTemplate = () => (`<section class="trip-main__trip-info  trip-info">
+</section>`);
