+++ src/view/trip-info.js
@@ -1,9 +1,6 @@
-import { formatDate } from '../utils';
 import { MOCK_EVENTS } from '../mock/events';
 
 const towns = MOCK_EVENTS.map((event) => event['destination'].name);
-
-
 const writePath = (places) => {
   let path = '';
 
@@ -23,13 +20,11 @@ const writePath = (places) => {
 };
 
 const writeDates = (dateFrom, dateTo) => {
-  const date = `${formatDate(dateFrom, 'MMM D')} — ${formatDate(dateTo, 'MMM D')}`;
+  const date = `${dateFrom.format('MMM D')} — ${dateTo.format('MMM D')}`;
   return date;
 };
 
-export const createTripInfoTemplate = () => (`
-	<div class="trip-info__main">
-    <h1 class="trip-info__title">${writePath(towns)}</h1>
-    <p class="trip-info__dates">${writeDates(MOCK_EVENTS[0]['date_from'], MOCK_EVENTS[MOCK_EVENTS.length -1]['date_to'])}</p>
-  </div>
-`);
+export const createTripInfoTemplate = () => (`<div class="trip-info__main">
+  <h1 class="trip-info__title">${writePath(towns)}</h1>
+  <p class="trip-info__dates">${writeDates(MOCK_EVENTS[0]['date_from'], MOCK_EVENTS[MOCK_EVENTS.length -1 ]['date_to'])}</p>
+</div>`);
